requirements.txt
django==6.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.11
django-cors-headers==4.4.0
gunicorn==21.2.0
drf-spectacular==0.27.0
drf-spectacular[sidecar]==0.27.0
django-ratelimit==4.1.0
django-filter==24.1
